Rename base->tvcm and put in thid_party/tvcm/src
In the continued effort to make the trace-viewer component model more
self-explanatory, the base namespace and thus the directory was renamed to
tvcm. The intent here is that we explain to contributors that tvcm is our
component model, which is a sort of strict layer above polymer. Then, anything
prefixed with tvcm is logically and obviously part of that model. Whereas,
prior to this patch, base/ and ui/ implied tvcm, but not in an obvious way.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1133 0e6d7f2b-9903-5b78-7403-59d27f066143