Remove parse_deps and move CalcLoadSequence to project
parse_deps was the original (and entirety) of tvcm. The new project, loader and
module factoring cleaned up a ton of the spaghetti that the parse_deps module
attempted to hide. Its one remaining value-add was that it put the implicit
dependency on base/__init__.js back into any attempt to load a module.  This
was fixed in a previous CL, rendering parse_deps completely obsolete.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1152 0e6d7f2b-9903-5b78-7403-59d27f066143