Add decoding of EventTrace event to get trace information.
The decoding of the first event (EventTrace:Header) is needed to detect
if the trace is a 32-bit or 64-bit version.

Originaly, this was propagate into the event header, but this rely on
code on the chrome side, which I pefer to keep as simple as possible.

R=nduca@chromium.org, nduca

BUG=

Review URL: https://codereview.appspot.com/70910043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1190 0e6d7f2b-9903-5b78-7403-59d27f066143