Move src->trace_viewer and build->trace_viewer/build
We need to use trace_viewer as a module to embed it in other projects.
This means that having a module laying around called "src" and "build" is
prone to conflicts (and does conflict, Android scripts for Chrome have a
build module). The solution is to put all trace_viewer python code that
is not related to tvcm into the trace_viewer folder, and add an __init__.py
so that it behaves exactly as one would expect.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1197 0e6d7f2b-9903-5b78-7403-59d27f066143