Introduce TimelineViewSidePanel
SidePanelContainer sits to the right of the regular timeline, and lets any part
of tracing add in a side-view to that panel. With followup patches, this can be
used to add domain-specific replacements or helper-tools for navigating through
the timeline. As part of this, had to clean up how selectors for the profiling and timeline view were written.

Collapsed: https://dl.dropboxusercontent.com/u/58819336/SidePanelExample.png

Expanded:
https://dl.dropboxusercontent.com/u/58819336/PanelExampleExpanded.png

Supports multiple panels, of course. :)

R=dsinclair@chromium.org, dsinclair
CC=miletus

Review URL: https://codereview.appspot.com/82990044

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1227 0e6d7f2b-9903-5b78-7403-59d27f066143