Highlight any flow events in the selection.
Prior to this CL there is no indication of a flow arc that has an event as
part of the selection. With this CL, I will add a yellow shadow to any flow
arc which has a start or end event in the selection.

Note, this is only one level of the arc. If there are 3 events in the flow
and only the last event is in the selection then only the last arc will
highlight. I'm not highlighting the entire flow arc for now.

https://docs.google.com/a/chromium.org/file/d/0BwBaQQ4lL3m2WUpoV1N3bmJ3bVE/preview

R=nduca@chromium.org

Review URL: https://codereview.appspot.com/107800043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1291 0e6d7f2b-9903-5b78-7403-59d27f066143