Add a simple record UI
This patch adds a simple record UI for the common recording cases.
Instead of a wall of checkboxes, the user gets a single dropdown box
with various presets (still need to adjust these), and the option to
manually adjust settings.

A screenshot is available at: http://pr.gg/simplerecord.png

More detailed changes:
1) An optional namespace parameter has been added to dom_helper.js's
createSelector. When creating the dropdown we now pass the
preset namespace so preset settings do not interfere with the
category settings.

2) The recordSelectionDialog_splitCategories test has been fixed to
set the settings key. This prevents collisions with other settings.

3) Most record_selection_dialog tests have been updated to use
manual settings mode so no presets are used.

4) When switching between manual and preset settings, the record
animates. This is the bulk of the style changes in this patch.

BUG=
R=nduca@chromium.org

Review URL: https://codereview.appspot.com/109040047

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1296 0e6d7f2b-9903-5b78-7403-59d27f066143