Show the preset trace categories
Existing tracers are unsure what categories are used for each of the
presets. This patch shows the categories behind each preset in a
light and small font. I think this is a nice compromise between our
existing trace experts and new users. A preview is at
http://pr.gg/tracecats.png

The complexity in this patch is just to keep the preset <-> manual
transition looking smooth.

One unrelated change is in this patch: instead of transitioning to
opacity:1, we stop at opacity:0.99 to prevent layer churn.

BUG=
R=dsinclair@chromium.org, nduca@chromium.org

Review URL: https://codereview.appspot.com/112140043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1337 0e6d7f2b-9903-5b78-7403-59d27f066143