Change how html files are decided to be modules
The first version of IsFilenameAModule used the presence of certain magical
tags in the filename as a signal that an html file was a tvcm module versus a
plain-old-html file.

The downside was that if a module was written without the proper tricks in it,
then it'd not be recognized. So, if you typed <!DOCTYPE HTML> instead of
<!DOCTYPE html> your module wouldn't be includable. And, other silly things
like that. This was really fragile.

This replaces the heuristics with a blacklist. All html files are assumed
to be modules, but then they can be added to the project blacklist for
non inclusion.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1383 0e6d7f2b-9903-5b78-7403-59d27f066143