Move tvcm/__init__.js to tvcm.html AND make it a dependency
In the old system, tvcm/__init__.js was brought in manually by
the bootstrap code. This isn't very component like. In this patch,
HTMLModule is updated to complain if tvcm is not somehow referenced by
a file containing tvcm.exportTo. This allows all bootstrap lines of
tvcm/__init__.js to be removed.

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1386 0e6d7f2b-9903-5b78-7403-59d27f066143