Fixed tab_view corner case issues.
Due to the asyncronous execution of the childrenUpdated_ call,
it is possible to have the selected property of the light-dom children
modified, before actually registering the child.

This patch fixes this problem.

Also, the clearSelectedTab_ was not actually clearing the selectedTab_
variable correctly, leading to a tab button staying selected.

R=nduca@chromium.org

Review URL: https://codereview.appspot.com/126850043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1452 0e6d7f2b-9903-5b78-7403-59d27f066143