Refactor Slice construction callsites to pass in args and duration.
This patch refactors Slice construction sites to pass in args and
duration parameters. This simplifies two AsyncSlice construction sites.
A more general solution of passing in startThread, endThread, and
sliceId is not possible because we depend on a common signature for
creating Slices (see SliceGroup::beginSlice). Adding optional arguments
to AsyncSlice is brittle because it tightly couples AsyncSlice and Slice.

R=nduca@chromium.org, nduca

Review URL: https://codereview.appspot.com/122320043

git-svn-id: http://trace-viewer.googlecode.com/svn/trunk@1453 0e6d7f2b-9903-5b78-7403-59d27f066143